Removed swift::storage::device::* defines
Previously, I had created a define intended to
wrap the creation of the storage device endpoints
with their associated ring configs.

I wound up decided against this for two reasons:
  - Abtstracting storage endpoints together with
the ring configurations can be confusing b/c it
conflates actions associated with the storage role
with actions associated with the ringbuilding role.
  - Abstracting the ring building hides side effects
of this manifest from the user (that it will effect
the ringbuilder role)

This pull request does the following:

- removes the swift::storage::device::* defines

- update the following code to use
swift::storage::device in combination with
ring_*_device
 -- examples/multi
 -- swift::storage::node