Renamed storage::device to storage::server
Changed the name of the defined resource type:
swift::storage::device to be swift::storage::server.

This change was made in order to better align with
the swift terminology since this define was
configuring a swift component server which has no
direct association with a device.

This commit also makes the swift::storage::server
the only place where the swift server template is
used. This allowed me to assume that the variables
device and bind_port will always be available.

Also updates declarations of the renamed type to
use the correct type name.

Removes the direct server file config from
swift::storage, replacing it instead with
declarations of swift::storage::server.