Add modue synlinks to gitignore
The spec tests now expect the modulepath
spec/fixtures/modules to contain links to all
of the modules that are required to run the
unit tests.

In order to run the tests, I need to create links
for these modules, but if I check in those links, it
will mess up the tests when they run in travis-ci.