Fix the rsync modules names
The rsync module name must be only ```$type``` as written in the
documentation [1].

The names with the port added exists but is intended to be used only in
SAIO mode [2] and is only supported if the ```vm_test_mode``` option is
set in ```swift.conf```.

The effect of these shares not being named correctly is seen in the
```swift-{object,account,container}-replicator``` daemons (replication
fails).

[1] http://swift.openstack.org/1.4.6/howto_installmultinode.html#configure-the-storage-nodes
[2] http://swift.openstack.org/1.4.6/development_saio.html