Drop Swift requirement on keystone:python.
I believe this requirement was added to obtain support auth_token
middleware (keystone auth support). This is now provided via
keystoneclient.

Since this requirement no longer works we should drop it since
it is heavyweight to install keystone source code on all the
swift proxies.

Most packages automatically require keystoneclient as well so
we can probably ignore automatically adding this dep in here. I'm open
to doing that if we thing it worthwhile... but we'll first need to
make a keystone::client module to support it.

Change-Id: I2a85ea964d2b9e1a93c3ec0ba8de31b188c140ea