Fix filter name for puppetdb
When creating the filter class name the last part is not capitalized.
So Swift::Storage::Filter::Healthcheck[bla] is named
Swift::Storage::Filter::healthcheck[bla].  Puppet don't detect the
syntax error as the resource is "created" within a string.  But
puppetdb's parser do not let this work.

This fix it and adjust the tests.

Change-Id: If11c3ad0e98d14b9387b3daa4afeb118387100f8
Closes-Bug: #1252998