Fix spurious warning in pipeline check
The storage::server class includes a bit of logic that checks
the pipeline and warns if ${type}-server isn't included.
However, the check has a logic flaw such that if $pipeline is an
array and contains ${type}-server, the warning is triggered
anyway.  This patch corrects the bad logic so that the warning
isn't triggered spuriously.

Change-Id: I7f81598f2f7910a2175fe772c9d82b2a1067826c
Closes-Bug: #1289187
(cherry picked from commit 01a3a40bcc43ffd492bea87e9b7b143bedccec37)