Update ssh module version
The ssh module had an issue with hostkeys on RHEL systems prior to
version 2.4.0 that would cause puppet to error during a run: the
sshdsakey fact returns empty, so the exported dsa sshkey resource fails
due to an empty parameter[1]. The current dependency restricts the ssh
module to less than 2.0, which contains the bug.

Problematically, version 2.5.0 of the ssh module adds a change that is
incompatible with older 3.x versions of puppet [2], which breaks our tests.

This patch updates the ssh dependency to use full range of the next
major version in order to be as accomodating as possible. However, for
RHEL users to avoid the ssh key bug, they must use at least version
2.4.0. For users of older puppet versions to avoid the bad parameter
issue, they must use a version less than 2.5.0. The .fixtures.yml pins
to 2.4.0 for maximum compatibility. We could pin the dependency in the
metadata to exactly 2.4.0 as well, but since these are issues in the
ssh module, not swift's use of the ssh module, it is more appropriate
just to allow the newer version but not be too restrictive on the
version.

The ssh::server class has an API change between 1.x and 2.x, but the
new parameters (ensure, storeconfigs_enabled, and an options hash) have
acceptable defaults.

[1] https://github.com/saz/puppet-ssh/blob/v1.4.0/manifests/hostkeys.pp#L8
[2] https://github.com/saz/puppet-ssh/blob/v2.5.0/manifests/server/install.pp#L7

Change-Id: I30216b0336b5f1f1c65e711cb97654ef663ba01f
(cherry picked from commit 7ad2a10e1fd5f78a6632cf76c8def3ae4559677e)