Add keystone to swift-proxy, use dispersion-report
The existing acceptence test configures keystone on the swift node
but does not configure the swift service to use keystone.  This change
adds the neccesary middleware to the proxy node to actually use keystone.
Further proof of keystone working is acheived by enabling the
swift-dispersion-report class which uploads test objects to swift
after first authenticating with keystone.  This change also updates the
call to keystone used by the dispersion class to an updated format.

Change-Id: Ibf97762f9b19c0ccb72178b04a1a86028bab8bcf