Fixed swift::proxy::cache requirement on memcached
Fixed the incorrect inclusion of a requirement for the memcached class
to be in the catalog for swift::proxy::cache due to a bad grep of
the memcached server list. The grep command will return an array of
matching values. Unfortunately this includes if nothing is found so the
if condition was always matching. Now if you are not using a local
memcache server on 127.0.0.1, the memcached class is not required to
be in the catalog.

Change-Id: I79b92dc67e46dfa0b62ba07b00633de6fb17bf1e
Closes-Bug: #1628967