Set concurrency to 1 for auditor/replicator/updater
Before this the default concurrency was set to $::processorcount, and
this resulted in environments where 24 auditors/replicators/reapers
processes were all using the same disk. This results in a high I/O load,
because every thread blocks the others.

These values are 1 by default in Swift. If one has many disks and a
powerful machine, a single-digit concurrency for auditors and
replicators might work fine; the reaper is most likely only running with
a concurrency of 1 in any case.

Let's set this to the default of 1; operators with a need to use a
higher value can easily override the defaults (and probably want to use
a value that is appropriate to their setup).

Closes-Bug: 1631359
Change-Id: I54a5d01e4bd357973e142274db79d119e9e46457