Add swift::storage::loopbacks class
This patch adds a wrapper for the storage::loopback definition, making
it possible to add multiple loopback devices using a Hiera hash.

Loopback devices are the preferred method if there is only a single
local disk with one partition, instead of just using a directory that is
created on the root filesystem.

This patch is very similar to I3692b17988f8eacb7b8a7df1006a7fe24f587065.

Change-Id: I11a230b7cf08a4cc2a144d9af0e6c81bb3827348