Change storage worker default to ::os_workers
This patch changes the default worker count from 1 to the ::os_workers fact.
::os_workers is based on the number of processors (currently cpu/4) but is
capped at a maximum of 8 worker processors.  This is a much more reasonable
default in general and prevents excessive resource consumption on systems with
a large number of CPUs.

This is a follow up to Ifa4f425091fae81204df68dd536f364efd647e3d

Related-Bug: 1655070
Change-Id: I29fd3d8930f32650549bfbf96478232075a7d0b9