[CVE-2016-9590] Correct configuration file perms
The configuration files for the swift proxy and storage server were
getting their permissions changed from 0640 to 0644 due to the missing
mode on the concat resource for these files. This change defines the
permissions for these files to be 0640 to prevent unauthorized reading
of sensitive information that may be contained in these files.

Depends-On: Icb466a9662ae34eedeb3fb65bcb618775cb97d5d
Change-Id: I67385d3a3f17f8baec657cc7f85153f9ff31bee4
(cherry picked from commit fff73b533b07e719b1daf65400dc47dcc84edf60)