Fix raising an error on rebalance warnings
If the ring has been rebalanced recently, the next call to rebalance
will not rebalance again if the time passed is shorter than
min_part_time. swift-ring-builder returns an exit code of 1 in this
case, which is a warning, but not a hard error (errors will use an exit
code of 2).

It makes sense therefore to expect an exit code of 0 or 1 and only fail
if it's different.

Closes-Bug: 1677588
Change-Id: I03dd075c590344a920099c57b6a7530f103e0ede