Configure www_authenticate_uri instead of auth_uri in authtoken
We're configuring a piece of keystonemiddleware, which saw this
change:
   https://review.openstack.org/508522

This minimal fix retains auth_uri wherever we were using it in
Puppet, because who cares? It's just a name of a class field.
But when we write it out, change it to the new name in the
configuration file, so these modules continue to work in the
future.

Depends-On: I7be3f7748a2857618dc4de38e09c7cccca6d2721
Change-Id: I63f8a4287235c4659807530851c917a6386b4dd8