Add servers_per_port parameter in object-server
According to the doc at this URL:
https://docs.openstack.org/swift/latest/deployment_guide.html

for performance reasons, it is desirable to use forks of the
object-server, each on a different port, one per device. This
is what I'm doing in openstack-cluster-install in Debian. For
this, I've been using this patch. It'd be nice to have it
upstream as well.

Change-Id: I9224eb7022871e92cc18f82e9985d91d5e0c2823