Add object_server_mb_per_sync
The default mb_per_sync value of Swift leads to adding a lot of
latencies when transfering to an object store. This patch adds
the possibility to override this value for object store (it is
less relevant for other backend which rarely will reach the
cache size anyway). According to our tests, a value of 16 MB
is very good in production.

Change-Id: Id379d4221bc7e2a0c16125b97140e39b7726fd02