Fix incorrect comments about updating region/zone of a device
Updating region/zone of a device is actually implemented and supported
in swift, while puppet-swift doesn't support that operation yet.
This patch updates error messages when an operator tries to update
region/zone of a device, so that they notice more clearly that the
cause exists in puppet-swift, not in swift.

Change-Id: I7c7f633630ac52578a0cf6beb2033465a16bd251