Fix disk mount options
The nobarrier option is invalid in recent kernels (>= 4.19), and having
it makes the mount fail. It has no effect on kernel >= 4.10.

Also, adding "nofail" to avoid failure when booting a server if
there is a broken drive, which is common on a large cluster.

Change-Id: Ib5262c26244363a1c47bdd562d560735ddc82a36
(cherry picked from commit 3c3a1212b802e9463b22b5941ccac5001056fde4)