Deprecate swift::test_file
This class generates a script to test swift installation but the script
has quite outdated implementation. It still required Keystone v2 API
which was already removed from Keystone, so no longer works.

Because verification of the deployment is out of scope of deployment
tools, the class is deprecated and will be removed in a future release.

Change-Id: I0409afde6d63d5c1b7b7193334545278faab4b0b