Multiple rsync modules
Currently with puppet-swift, each server has a global [ object ] rsync
module name. As a consequence, the number of rsync connection is set
globally for each server.

With this patch, it's possible to define one rsync module per device,
which makes is to much better: it handles both the cases where a hard
drive fails, or when all drives need to sync data.

Note that with this patch, the default is to keep the old behavior of
one single rsync module per service type (a / c / o).

Change-Id: Id2ad5b10b9286c3740d6ebdbc61d2fd9792ba496