Remove logic for old swift-ring-builder
swift 2.10.0 was released in 2016. We no longer need to support even
older versions. The output format depends on the version currently
installed, thus the latest logic should work as long as the same
release (or even a bit older release) is installed.

This also updates the logic to skip the header lines. The new logic
is considered to be more robust, because it works even when additional
lines are added.

Change-Id: I694ca4c046f490e4ffecf2101f3f147a69045162