Fix ignored notifications/orders
This fixes the resource notifications and orders defined in the mount
class was unintentionally ignored unless selinux is enabled.

Also, resource collectors are replaced by explicit resources, because
we can expect precise list of resources defined in the class.

Finally, restorecon should be executed when the directory is mounted,
regardless of the result of chmod.

Change-Id: If919c7aceb5394a99657c3be7dc4f376d8a1b453