container-reconciler: Enable cache middleware by default
The chache middleware is now enabled by default in most of the services
but only container-reconciler disables it by default, though the sample
config container-reconciler config file in the swift repository enables
the middleware.

This updates the default value so that the default pipeline is
consistent among services, and is aligned with the sample files in
swift.

Change-Id: Ifcb4d241de14a9e0b7e3afb79e73b9a57acc9d9e