Fix constraint related to dnspython
The title of the package resource to manage dnspython is not python3-*
but python-*. Also the proxy-server service should be restarted when
the package is updated so that the updated version is loaded.

Backport note:
Because the swift-proxy-server tag is not used in stable/2023.1, this
modifies the way to search the swift-proxy-server service resource.

Change-Id: I3372c38dcbc7312395ddf643862da91b00262ad7
(cherry picked from commit f77b2c23c9407a7aa4c434959470b5b8deb95bcd)
(cherry picked from commit 134bce281286db7c3c772e179e2b1da2ea459e28)