Fix handling of cors options
The strict_cors_mode option should be set independently from
the cors_allow_origin, for example in case when strict mode is
disabled.

Also the current description about the strict_cors_mode is completely
incorrect.

This fixes these problems, and also allows usage of an array value
for cors_allow_origin so that users can compose a value to be passed
to the parameter more easily.

NOTE:
The strict_cors_mode option defaults to true in swift, thus its default
is replaced by the os_service_default fact.

Closes-Bug: #2048840
Change-Id: I59317519e31349d0b0b7efafae3a0a8a44d81d37