Per daemon log name
Currently, puppet-swift can only set log_name. However, it's possible
with swift to have each daemon log with a different name, making it
possible to have logging for each daemon in a separate file, which is
often a way more convenient for operators.

This patch sets a different log_name for each daemon if set to true.
The default behavior is left unchanged.

Change-Id: I27386f6ac00b2efcfa5fc4f5ed3b86efc49c17fc